requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
anthropic
python-dotenv
huggingface_hub
google-generativeai
google-genai
chromadb
sentence-transformers
langchain